+++ client/private/js/src/constants.js
@@ -11,6 +11,6 @@
 // const SERVER_ROOT_PATH = 'https://www.shivayl.com/';
 const API_ROOT_PATH = `${SERVER_ROOT_PATH}portfolio-os/api/`;
 const IMG_PATH = `${SERVER_ROOT_PATH}img/`;
+const DEFAULT_APP_ICON = `${IMG_PATH}default-taskbar-icon-white.svg`;
 const AUTH_TOKEN_ID = 'JWT';
-const DEFAULT_APP_ICON = '';
 const START_MENU_ANIM_DELAY = 1;