+++ client/private/js/src/constants.js
@@ -14,3 +14,5 @@ const IMG_PATH = `${SERVER_ROOT_PATH}img/`;
 const DEFAULT_APP_ICON = `${IMG_PATH}default-taskbar-icon-white.svg`;
 const AUTH_TOKEN_ID = 'JWT';
 const START_MENU_ANIM_DELAY = 1;
+
+const ERROR_MSG_INSTALL_APP = 'There was an error while installing the app.';