+++ server/portfolio-os/routes/users.js
@@ -7,6 +7,8 @@
  *
  */
 
+// api/portfolio-os/(users | user )/
+
 'use strict';
 const userStore = require( '../dataAccess/userStore' );
 const { sanitizeHTML } = require( '../../../common/commonUtils' );
@@ -165,7 +167,11 @@ module.exports = {
 
   installApp: async ( req, res ) => {
     try {
-      console.log( 'install app:', req.body );
+      const affectedRows = await userStore.installApp( req.user.id, req.params.appId );
+      if ( affectedRows <= 0 )
+        return res.status( 500 ).json( 0 );
+
+      return res.status( 200 ).json( affectedRows );
 
     } catch ( e ) {
       return res.status( 500 ).json( -1 );